apps/identity/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "identity"
version = "0.1.0"
description = "Rule-based deterministic matching service"
requires-python = ">=3.11"
dependencies = [
    "edna-common",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]

